refactor(tui): merge model and view into statusview.go
Follow codebase conventions (e.g. sessionform.go) — single file for
the model+view rather than splitting across model.go and view.go.
Renamed to statusview.go to match the package name.
