+++ internal/tui/statusview/statusview.go
@@ -10,6 +10,7 @@ import (
 	"github.com/charmbracelet/bubbles/help"
 	"github.com/charmbracelet/bubbles/key"
 	tea "github.com/charmbracelet/bubbletea"
+	"github.com/charmbracelet/lipgloss"
 	"github.com/eleonorayaya/utena/internal/session"
 	"github.com/eleonorayaya/utena/internal/tmux"
 	"github.com/eleonorayaya/utena/internal/tui/provider"
@@ -154,6 +155,47 @@ func (m *Model) syncTabs() {
 	}
 }
 
+func (m Model) View() string {
+	if m.isExpanded() {
+		return m.expandedView()
+	}
+	return m.collapsedView()
+}
+
+func (m Model) collapsedView() string {
+	maxNameLen := m.width - 2
+	if maxNameLen < 1 {
+		maxNameLen = 1
+	}
+
+	var lines []string
+	for _, sess := range m.orderedActiveSessions() {
+		name := sessionDisplayName(sess)
+		name = truncate(name, maxNameLen)
+		lines = append(lines, "  "+name)
+	}
+
+	return strings.Join(lines, "\n")
+}
+
+func (m Model) expandedView() string {
+	ordered := m.orderedActiveSessions()
+
+	var parts []string
+	for _, s := range ordered {
+		if tab, ok := m.tabs[s.ID]; ok {
+			parts = append(parts, tab.View())
+		}
+	}
+
+	if len(parts) == 0 {
+		empty := lipgloss.NewStyle().Foreground(colorTextMuted)
+		return empty.Render("  no active sessions")
+	}
+
+	return strings.Join(parts, "\n")
+}
+
 func (m Model) focusNextPane() {
 	if m.paneID == "" {
 		return
@@ -211,4 +253,21 @@ func (m Model) orderedActiveSessions() []session.Session {
 		return sessionDisplayName(result[i]) < sessionDisplayName(result[j])
 	})
 	return result
+}
+
+func sessionDisplayName(s session.Session) string {
+	if s.Name != "" {
+		return s.Name
+	}
+	return s.TmuxSessionName
+}
+
+func truncate(s string, maxLen int) string {
+	if len(s) <= maxLen {
+		return s
+	}
+	if maxLen <= 1 {
+		return s[:maxLen]
+	}
+	return s[:maxLen-1] + "…"
 }
